wasmparser: improve visit_operator performance
- The BinaryReader can visit_operator by itself again,
  with no need for OperatorsReader or extra state, when
  the visitor implements the FrameStack trait. Validation
  uses this to avoid duplicating the frame stack.

- OperatorsReader keeps an internal FrameStack and can be created
  from its allocations (and turned back into them), avoiding
  the need to allocate a stack of frames.
