+++ crates/wasmparser/src/validator/func.rs
@@ -1,6 +1,6 @@
 use super::operators::{Frame, OperatorValidator, OperatorValidatorAllocations};
-use crate::{BinaryReader, OperatorsReader, Result, ValType, VisitOperator};
-use crate::{FunctionBody, ModuleArity, Operator, WasmFeatures, WasmModuleResources};
+use crate::{BinaryReader, Result, ValType, VisitOperator};
+use crate::{FrameStack, FunctionBody, ModuleArity, Operator, WasmFeatures, WasmModuleResources};
 
 /// Resources necessary to perform validation of a function.
 ///
@@ -84,24 +84,27 @@ impl<T: WasmModuleResources> FuncValidator<T> {
         {
             reader.set_features(self.validator.features);
         }
-        let mut ops = OperatorsReader::new(reader);
-        while !ops.eof() {
+        while !reader.eof() {
             // In a debug build, verify that the validator's pops and pushes to and from
             // the operand stack match the operator's arity.
             #[cfg(debug_assertions)]
             let (pop_push_snapshot, arity) = (
                 self.validator.pop_push_count,
-                ops.clone()
-                    .read()?
-                    .operator_arity(&self.visitor(ops.original_position())),
+                (|| {
+                    let operator = reader
+                        .clone()
+                        .peek_operator(&self.visitor(reader.original_position()))
+                        .ok()?;
+                    operator.operator_arity(&self.visitor(reader.original_position()))
+                })(),
             );
 
-            ops.visit_operator(&mut self.visitor(ops.original_position()))??;
+            reader.visit_operator(&mut self.visitor(reader.original_position()))??;
 
             #[cfg(debug_assertions)]
             {
                 let (params, results) = arity.ok_or(format_err!(
-                    ops.original_position(),
+                    reader.original_position(),
                     "could not calculate operator arity"
                 ))?;
 
@@ -113,7 +116,7 @@ impl<T: WasmModuleResources> FuncValidator<T> {
                 }
             }
         }
-        ops.finish()
+        reader.finish_expression(&self.visitor(reader.original_position()))
     }
 
     /// Reads the local definitions from the given `BinaryReader`, often sourced
@@ -161,18 +164,18 @@ impl<T: WasmModuleResources> FuncValidator<T> {
     /// pub fn validate<R>(validator: &mut FuncValidator<R>, body: &FunctionBody<'_>) -> Result<()>
     /// where R: WasmModuleResources
     /// {
-    ///     let mut operator_reader = body.get_operators_reader()?;
+    ///     let mut operator_reader = body.get_binary_reader_for_operators()?;
     ///     while !operator_reader.eof() {
     ///         let mut visitor = validator.visitor(operator_reader.original_position());
     ///         operator_reader.visit_operator(&mut visitor)??;
     ///     }
-    ///     operator_reader.finish()
+    ///     operator_reader.finish_expression(&validator.visitor(operator_reader.original_position()))
     /// }
     /// ```
     pub fn visitor<'this, 'a: 'this>(
         &'this mut self,
         offset: usize,
-    ) -> impl VisitOperator<'a, Output = Result<()>> + ModuleArity + 'this {
+    ) -> impl VisitOperator<'a, Output = Result<()>> + ModuleArity + FrameStack + 'this {
         self.validator.with_resources(&self.resources, offset)
     }
